Move capi models into commercial.model.capi.
