+++ commercial/app/controllers/TrafficDriverController.scala
@@ -1,14 +1,15 @@
 package commercial.controllers
 
+import commercial.model.capi.CapiAgent
 import common.{Edition, ExecutionContexts, JsonComponent, Logging}
 import contentapi.ContentApiClient
-import model.commercial.{CapiAgent, TrafficDriver}
+import model.commercial.TrafficDriver
 import model.{Cached, ContentType}
 import play.api.mvc.{Action, AnyContent, Controller, Request}
 
 import scala.concurrent.Future
-import scala.util.control.NonFatal
 import scala.concurrent.duration.DurationInt
+import scala.util.control.NonFatal
 
 class TrafficDriverController(
     contentApiClient: ContentApiClient,
